将数据库中的tinyint 对应byte 将CREATE_TIME修改为CREATION_TIME 将constant包中的常量整理到pojo

+++ src/main/java/cn/com/wh/ring/app/dao/user/UserDao.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 public interface UserDao {
     void insert(User user);
 
-    void updatePassword(@Param("account") String account, @Param("accountType") int accountType, @Param("password") String password);
+    void updatePassword(@Param("account") String account, @Param("accountType") byte accountType, @Param("password") String password);
 
     void updateToken(User user);
 
@@ -21,7 +21,7 @@ public interface UserDao {
 
     Long queryMaxUserId();
 
-    User queryByAccount(@Param("account") String account, @Param("accountType") int accountType);
+    User queryByAccount(@Param("account") String account, @Param("accountType") byte accountType);
 
     User queryByBinAccount(@Param("account") String account);
